Initial stream and source services
This commit introduces the stream server and source server. The stream
server handles incoming stream requests, including requests to record
video.

The camera providers will issue a `record` RPC call and get back a
session. The session information will tell them where to stream their
video and audio.

Camera players (extensions that can stream video) and extensions that
need to stream video from events both call the stream API to set up a
stream session which includes the RPC address and ports to send the
video and audio.

The source server handles mapping if a source is from an event or from a
camera directly.